chore: rename core pkg to flect
