+++ internal/flect/core.go
@@ -1,4 +1,4 @@
-package core
+package flect
 
 import (
 	"strings"
