+++ internal/flect/custom_data.go
@@ -1,4 +1,4 @@
-package core
+package flect
 
 import (
 	"bytes"
