+++ internal/flect/plural_rules.go
@@ -1,4 +1,4 @@
-package core
+package flect
 
 import "fmt"
