+++ internal/flect/singular_rules.go
@@ -1,4 +1,4 @@
-package core
+package flect
 
 var singularRules = []rule{}
